+++ inst/include/Rcpp/complex.h
@@ -34,24 +34,25 @@ inline Rcomplex operator/( const Rcomplex& a, const Rcomplex& b){
     double ratio, den;
     double abr, abi;
 
-    if( (abr = b.r) < 0)
-    abr = - abr;
-    if( (abi = b.i) < 0)
-    abi = - abi;
+    if( (abr = b.r) < 0) abr = - abr;
+    if( (abi = b.i) < 0) abi = - abi;
     if( abr <= abi ) {
-    ratio = b.r / b.i ;
-    den = b.i * (1 + ratio*ratio);
-    c.r = (a.r*ratio + a.i) / den;
-    c.i = (a.i*ratio - a.r) / den;
+        ratio = b.r / b.i ;
+        den = b.i * (1 + ratio*ratio);
+        c.r = (a.r*ratio + a.i) / den;
+        c.i = (a.i*ratio - a.r) / den;
+    } else {
+        ratio = b.i / b.r ;
+        den = b.r * (1 + ratio*ratio);
+        c.r = (a.r + a.i*ratio) / den;
+        c.i = (a.i - a.r*ratio) / den;
     }
-    else {
-    ratio = b.i / b.r ;
-    den = b.r * (1 + ratio*ratio);
-    c.r = (a.r + a.i*ratio) / den;
-    c.i = (a.i - a.r*ratio) / den;
-    }
-    return c ;
+    return c ;  
 }
+inline Rcomplex operator/( Rcomplex a, double n){
+    return Rcomplex{ a.r / n, a.i / n } ;
+}
+
 inline bool operator==( const Rcomplex& a, const Rcomplex& b){
     return a.r == b.r && a.i == b.i ;    
 }